\name{@PKGNAME@}
\docType{data}
\alias{@PKGNAME@}
\title{Vectors used by fRMA for @ARRAYTYPE@.}
\description{This data object was automatically created by the @CREATOR@.}
\usage{data(@PKGNAME@)}
\format{A list with 5 elements.
  \tabular{ll}{
    normVec \tab normalization vector \cr
    probeVec \tab probe effect vector \cr
    probeVarWithin \tab within batch probe variance \cr
    probeVarBetween \tab between batch probe variance \cr
    probesetSD \tab within probeset standard deviation \cr
}}

\keyword{datasets}
\examples{
 data(@PKGNAME@) 
 str(@PKGNAME@)
}
